pyproject.toml
[tool.poetry]
name = "world-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
supabase = "^2.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"